Use new API for pipeline astrometry plots
