+++ drizzlepac/haputils/quality_analysis.py
@@ -42,18 +42,9 @@ function myFunction() {
 
 """
 import json
-import os
 import sys
 from datetime import datetime
 import time
-
-from drizzlepac.haputils.graph_utils import HAPFigure
-import drizzlepac.haputils.graph_utils as gu
-
-from bokeh.layouts import row, column
-from bokeh.plotting import figure, output_file, save, show
-from bokeh.models import ColumnDataSource, Label, Range1d
-from bokeh.models.tools import HoverTool
 
 from astropy.table import Table, vstack
 from astropy.io import fits
@@ -66,9 +57,7 @@ from stsci.tools import logutil
 import tweakwcs
 
 from . import astrometric_utils as amutils
-from .. import tweakutils
 from . import diagnostic_utils as du
-from .pandas_utils import PandasDFReader
 
 
 MSG_DATEFMT = '%Y%j%H%M%S'
@@ -423,569 +412,4 @@ def run_all(input, files, log_level=logutil.logging.NOTSET):
     print("Generated quality statistics as {}".format(json_files))
 
 
-# -------------------------------------------------------------------------------
-#
-#  Code for generating relevant plots from these results
-#
-# -------------------------------------------------------------------------------
-def generate_plots(json_data):
-    """Create plots from json file or json data"""
 
-    if isinstance(json_data, str):
-        # Open json file and read in data
-        with open(json_data) as jfile:
-            json_data = json.load(jfile)
-
-    fig_id = 0
-    for fname in json_data:
-        data = json_data[fname]
-
-        rootname = fname.split("_")[0]
-        coldata = [data['x'], data['y'], data['ref_x'], data['ref_y']]
-        # Insure all columns are numpy arrays
-        coldata = [np.array(c) for c in coldata]
-        title_str = 'Residuals\ for\ {0}\ using\ {1:6d}\ sources'.format(
-                    fname.replace('_', '\_'), data['nmatches'])
-
-        vector_name = '{}_vector_quality.png'.format(rootname)
-        resids_name = '{}_resids_quality.png'.format(rootname)
-        # Generate plots
-        tweakutils.make_vector_plot(None, data=coldata,
-                                    figure_id=fig_id, title=title_str, vector=True,
-                                    plotname=vector_name)
-        fig_id += 1
-        tweakutils.make_vector_plot(None, data=coldata, ylimit=0.5,
-                                    figure_id=fig_id, title=title_str, vector=False,
-                                    plotname=resids_name)
-        fig_id += 1
-
-
-# -------------------------------------------------------------------------------
-# Generate the Bokeh plot for the pipeline astrometric data.
-#
-HOVER_COLUMNS = ['gen_info.imgname',
-                 'gen_info.instrument',
-                 'gen_info.detector',
-                 'gen_info.filter',
-                 'header.DATE-OBS',
-                 'header.RA_TARG',
-                 'header.DEC_TARG',
-                 'header.GYROMODE',
-                 'header.EXPTIME',
-                 'fit_results.aligned_to']
-
-TOOLTIPS_LIST = ['EXPNAME', 'INSTRUMENT', 'DET', 'FILTER',
-                 'DATE', 'RA', 'DEC', 'GYRO', 'EXPTIME',
-                 'ALIGNED_TO']
-INSTRUMENT_COLUMN = 'full_instrument'
-
-RESULTS_COLUMNS = ['fit_results.rms_x',
-                   'fit_results.rms_y',
-                   'fit_results.xsh',
-                   'fit_results.ysh',
-                   'fit_results.rot',
-                   'fit_results.scale',
-                   'fit_results.rot_fit',
-                   'fit_results.scale_fit',
-                   'fit_results.nmatches',
-                   'fit_results.skew']
-
-RESIDS_COLUMNS = ['residuals.x',
-                  'residuals.y',
-                  'residuals.ref_x',
-                  'residuals.ref_y']
-TOOLSEP_START = '{'
-TOOLSEP_END = '}'
-DETECTOR_LEGEND = {'UVIS': 'magenta', 'IR': 'red', 'WFC': 'blue',
-                   'SBC': 'yellow', 'HRC': 'black'}
-
-FIGURE_TOOLS = 'pan,wheel_zoom,box_zoom,zoom_in,zoom_out,box_select,undo,reset,save'
-"""
-This code comes from:
-
- https://github.com/holoviz/holoviews/blob/master/holoviews/plotting/bokeh/chart.py#L309
-
-This code shows how they define the arrow heads for the vector plots.  A
-'non-invasive' way to use the Bokeh Segment Glyph to create a vector plot would
-be to define the arrow heads as 2 additional segments starting at the endpoint
-of the vector. The new segments would then be added to the data points to be
-plotted by Segment to appear as an arrow head on each segment.
-
-        # Get x, y, angle, magnitude and color data
-        rads = element.dimension_values(2)
-        if self.invert_axes:
-            xidx, yidx = (1, 0)
-            rads = np.pi/2 - rads
-        else:
-            xidx, yidx = (0, 1)
-        lens = self._get_lengths(element, ranges)/input_scale
-        cdim = element.get_dimension(self.color_index)
-        cdata, cmapping = self._get_color_data(element, ranges, style,
-                                               name='line_color')
-
-        # Compute segments and arrowheads
-        xs = element.dimension_values(xidx)
-        ys = element.dimension_values(yidx)
-
-        # Compute offset depending on pivot option
-        xoffsets = np.cos(rads)*lens/2.
-        yoffsets = np.sin(rads)*lens/2.
-        if self.pivot == 'mid':
-            nxoff, pxoff = xoffsets, xoffsets
-            nyoff, pyoff = yoffsets, yoffsets
-        elif self.pivot == 'tip':
-            nxoff, pxoff = 0, xoffsets*2
-            nyoff, pyoff = 0, yoffsets*2
-        elif self.pivot == 'tail':
-            nxoff, pxoff = xoffsets*2, 0
-            nyoff, pyoff = yoffsets*2, 0
-        x0s, x1s = (xs + nxoff, xs - pxoff)
-        y0s, y1s = (ys + nyoff, ys - pyoff)
-
-        color = None
-        if self.arrow_heads:
-            arrow_len = (lens/4.)
-            xa1s = x0s - np.cos(rads+np.pi/4)*arrow_len
-            ya1s = y0s - np.sin(rads+np.pi/4)*arrow_len
-            xa2s = x0s - np.cos(rads-np.pi/4)*arrow_len
-            ya2s = y0s - np.sin(rads-np.pi/4)*arrow_len
-            x0s = np.tile(x0s, 3)
-            x1s = np.concatenate([x1s, xa1s, xa2s])
-            y0s = np.tile(y0s, 3)
-            y1s = np.concatenate([y1s, ya1s, ya2s])
-            if cdim and cdim.name in cdata:
-                color = np.tile(cdata[cdim.name], 3)
-        elif cdim:
-            color = cdata.get(cdim.name)
-
-        data = {'x0': x0s, 'x1': x1s, 'y0': y0s, 'y1': y1s}
-        mapping = dict(x0='x0', x1='x1', y0='y0', y1='y1')
-"""
-
-
-# MDD Deprecated
-# This is only still here if you want to run the graphics in this routine to see
-# what the code does.  This routine has already been migrated in a mildly changed
-# form to graph_utils.py.
-def build_tooltips(tips):
-    """Return list of tuples for tooltips to use in hover tool.
-
-    Parameters
-    ----------
-    tips : list
-        List of indices for the HOVER_COLUMNS entries to be used as tooltips
-        to be included in the hover tool.
-
-    """
-    tools = [(TOOLTIPS_LIST[i], '@{}{}{}'.format(
-                                TOOLSEP_START,
-                                HOVER_COLUMNS[i],
-                                TOOLSEP_END)) for i in tips]
-
-    return tools
-
-
-# *** We should think about having a more general get_pandas_data for
-# both astrometry and svm graphics. ***
-# MDD Deprecated maybe - This routine returns two dataframes so this may
-# be desired for the "astrometry" routines which should continue to live in
-# this module. 
-# NOTE: The returned value from the
-# migrated version of get_pandas_data is a DataFrame
-# and not a ColumnDataSource.   I suggest you do not convert the DataFrame
-# into a ColumnDataSource until you are ready to plot as this is really
-# what the ColumnDataSource is there to support (IMHO).
-def get_pandas_data(pandas_filename):
-    """Load the harvested data, stored in a CSV file, into local arrays.
-
-    Parameters
-    ==========
-    pandas_filename: str
-        Name of the CSV file created by the harvester.
-
-    Returns
-    =======
-    phot_data: Pandas dataframe
-        Dataframe which is a subset of the input Pandas dataframe written out as
-        a CSV file.  The subset dataframe consists of only the requested columns
-        and rows where all of the requested columns did not contain NaNs.
-
-    """
-
-    # Instantiate a Pandas Dataframe Reader (lazy instantiation)
-    # df_handle = PandasDFReader_CSV("svm_qa_dataframe.csv")
-    df_handle = PandasDFReader(pandas_filename, log_level=logutil.logging.NOTSET)
-
-    # In this particular case, the names of the desired columns do not
-    # have to be further manipulated, for example, to add dataset specific
-    # names.
-    #
-    # Get the relevant column data, eliminating all rows which have NaNs
-    # in any of the relevant columns.
-    if pandas_filename.endswith('.h5'):
-        fit_data = df_handle.get_columns_HDF5(HOVER_COLUMNS + RESULTS_COLUMNS)
-        resids_data = df_handle.get_columns_HDF5(RESIDS_COLUMNS)
-    else:
-        fit_data = df_handle.get_columns_CSV(HOVER_COLUMNS + RESULTS_COLUMNS)
-        resids_data = df_handle.get_columns_CSV(RESIDS_COLUMNS)
-
-    return fit_data, resids_data
-
-
-# MDD Deprecated - When generate_summary_plots() is upgraded. This code can be deleted.
-def build_circle_plot(**plot_dict):
-    """Create figure object for plotting desired columns as a scatter plot with circles
-
-    Parameters
-    ----------
-    source : Pandas ColumnDataSource
-        Object with all the input data
-
-    x, y : str
-        Names of X and Y columns of data from data `source`
-
-    x_label, y_label : str
-        Labels to use for the X and Y axes (respectively)
-
-    title : str
-        Title of the plot
-
-    tips : list of int
-        List of indices for the columns from `source` to use as hints
-        in the HoverTool
-
-    color : string, optional
-        Single color to use for data points in the plot if `colormap` is not used
-
-    colormap : bool, optional
-        Specify whether or not to use a pre-defined set of colors for the points
-        derived from the `colormap` column in the input data `source`
-
-    markersize : int, optional
-        Size of each marker in the plot
-
-    legend_group : str, optional
-        If `colormap` is used, this is the name of the column from the input
-        data `source` to use for defining the legend for the colors used.  The
-        same colors should be assigned to all the same values of data from the
-        column, for example, all 'ACS/WFC' data points from the `instrument`
-        column should have a `colormap` column value of `blue`.
-
-    """
-    # Interpret required elements
-    x = plot_dict['x']
-    y = plot_dict['y']
-    title = plot_dict['title']
-    x_label = plot_dict['x_label']
-    y_label = plot_dict['y_label']
-    tips = plot_dict['tips']
-    source = plot_dict['source']
-
-    # check for optional elements
-    color = plot_dict.get('color', 'blue')
-    click_policy = plot_dict.get('click_policy', 'hide')
-    markersize = plot_dict.get('markersize', 10)
-    colormap = plot_dict.get('colormap', False)
-    legend_group = plot_dict.get('legend_group')
-
-    # Define a figure object
-    p1 = figure(tools=FIGURE_TOOLS)
-
-    if colormap:
-        # Add the glyphs
-        # This will use the 'colormap' column from 'source' for the colors of
-        # each point.  This column should have been populated by the calling
-        # routine.
-        p1.circle(x=x, y=y, source=source,
-                  fill_alpha=0.5, line_alpha=0.5,
-                  size=markersize, color='colormap',
-                  legend_group=legend_group)
-    else:
-        # Add the glyphs
-        p1.circle(x=x, y=y, source=source,
-                  fill_alpha=0.5, line_alpha=0.5,
-                  size=markersize)
-
-    p1.legend.click_policy = click_policy
-
-    p1.title.text = title
-    p1.xaxis.axis_label = x_label
-    p1.yaxis.axis_label = y_label
-
-    hover_p1 = HoverTool()
-    tools = build_tooltips(tips)
-    hover_p1.tooltips = tools
-    p1.add_tools(hover_p1)
-
-    return p1
-
-
-# This routine has been fully upgraded..
-def build_vector_plot(sourceCDS, **plot_dict):
-    """Create figure object for plotting desired columns as a scatter plot with circles
-
-    Parameters
-    ----------
-    sourceCDS : Pandas ColumnDataSource
-        Object with all the input data
-
-    x_label, y_label : str, optional
-        Labels to use for the X and Y axes (respectively)
-
-    title : str, optional
-        Title of the plot
-
-    color : string, optional
-        Single color to use for data points in the plot if `colormap` is not used
-
-    """
-    # Interpret required elements
-    sourceCDS = sourceCDS
-
-    # Check for optional elements
-    x_label = plot_dict.get('x_label', '')
-    y_label = plot_dict.get('y_label', '')
-    title = plot_dict.get('title', '')
-    color = plot_dict.get('color', 'blue')
-
-    # Define a figure object with square aspect ratio
-    p1 = HAPFigure(title=title,
-                   x_label=x_label,
-                   y_label=y_label,
-                   use_hover_tips=False)
-
-    # Add the glyphs
-    p1.build_vector_glyph(sourceCDS,
-                          color=color)
-
-    return p1
-
-
-# This routine has not been touched and needs to be upgraded.
-def generate_summary_plots(fitCDS, display_plot, output='cal_qa_results.html'):
-    """Generate the graphics associated with this particular type of data.
-
-    Parameters
-    ==========
-    fitCDS : ColumnDataSource object
-        Dataframe consisting of the relative alignment astrometric fit results
-
-    display_plot : bool, optional
-        Option to display the plot to the screen
-        Default: False
-
-    output : str, optional
-        Filename for output file with generated plot
-
-    Returns
-    -------
-    output : str
-        Name of HTML file where the plot was saved.
-
-    NOTES
-    -----
-    Example from bokeh.org on how to create a tabbed set of plots:
-
-    .. code:: python
-
-        from bokeh.io import output_file, show
-        from bokeh.models import Panel, Tabs
-        from bokeh.plotting import figure
-
-        output_file("slider.html")
-
-        p1 = figure(plot_width=300, plot_height=300)
-        p1.circle([1, 2, 3, 4, 5], [6, 7, 2, 4, 5], size=20, color="navy", alpha=0.5)
-        tab1 = Panel(child=p1, title="circle")
-
-        p2 = figure(plot_width=300, plot_height=300)
-        p2.line([1, 2, 3, 4, 5], [6, 7, 2, 4, 5], line_width=3, color="navy", alpha=0.5)
-        tab2 = Panel(child=p2, title="line")
-
-        tabs = Tabs(tabs=[ tab1, tab2 ])
-
-        show(tabs)
-
-
-    """
-    # TODO: include the date from the input data as part of the html filename
-    # Set the output file immediately as advised by Bokeh.
-    output_file(output)
-
-    num_of_datasets = len(fitCDS.data['index'])
-    print('Number of datasets: {}'.format(num_of_datasets))
-
-    colormap = [DETECTOR_LEGEND[x] for x in fitCDS.data[HOVER_COLUMNS[2]]]
-    fitCDS.data['colormap'] = colormap
-    inst_det = ["{}/{}".format(i, d) for (i, d) in zip(fitCDS.data[HOVER_COLUMNS[1]],
-                                                       fitCDS.data[HOVER_COLUMNS[2]])]
-    fitCDS.data[INSTRUMENT_COLUMN] = inst_det
-
-    plot_list = []
-
-    p1 = [build_circle_plot(x=RESULTS_COLUMNS[0], y=RESULTS_COLUMNS[1],
-                            source=fitCDS,
-                            title='RMS Values',
-                            x_label="RMS_X (pixels)",
-                            y_label='RMS_Y (pixels)',
-                            tips=[0, 1, 2, 3, 8],
-                            colormap=True, legend_group=INSTRUMENT_COLUMN)]
-    plot_list += p1
-
-    p2 = [build_circle_plot(x=RESULTS_COLUMNS[2], y=RESULTS_COLUMNS[3],
-                            source=fitCDS,
-                            title='Offsets',
-                            x_label="SHIFT X (pixels)",
-                            y_label='SHIFT Y (pixels)',
-                            tips=[0, 1, 2, 3, 8],
-                            colormap=True, legend_group=INSTRUMENT_COLUMN)]
-    plot_list += p2
-
-    p3 = [build_circle_plot(x=RESULTS_COLUMNS[8], y=RESULTS_COLUMNS[4],
-                            source=fitCDS,
-                            title='Rotation',
-                            x_label="Number of matched sources",
-                            y_label='Rotation (degrees)',
-                            tips=[0, 1, 2, 3, 8],
-                            colormap=True, legend_group=INSTRUMENT_COLUMN)]
-    plot_list += p3
-
-    p4 = [build_circle_plot(x=RESULTS_COLUMNS[8], y=RESULTS_COLUMNS[5],
-                            source=fitCDS,
-                            title='Scale',
-                            x_label="Number of matched sources",
-                            y_label='Scale',
-                            tips=[0, 1, 2, 3, 8],
-                            colormap=True, legend_group=INSTRUMENT_COLUMN)]
-    plot_list += p4
-
-    p5 = [build_circle_plot(x=RESULTS_COLUMNS[8], y=RESULTS_COLUMNS[9],
-                            source=fitCDS,
-                            title='Skew',
-                            x_label="Number of matched sources",
-                            y_label='Skew (degrees)',
-                            tips=[0, 1, 2, 3, 8],
-                            colormap=True, legend_group=INSTRUMENT_COLUMN)]
-    plot_list += p5
-
-    # Save the generated plots to an HTML file define using 'output_file()'
-    # Display and save
-    if display_plot:
-        show(column(plot_list))
-    else:
-        print('Saving: {}'.format(output))
-        save(column(plot_list))
-
-    return output
-
-
-# This is a fully upgraded function. 
-def generate_residual_plots(residsCDS, filename, display_plot, output_dir=None, output=''):
-    rootname = '_'.join(filename.split("_")[:-1])
-    output = '{}_vectors_{}'.format(rootname, output)
-    if output_dir is not None:
-        output = os.path.join(output_dir, output)
-    output_file(output)
-
-    delta_x = np.array(residsCDS.data['x']) - np.array(residsCDS.data['xr'])
-    delta_y = np.array(residsCDS.data['y']) - np.array(residsCDS.data['yr'])
-    residsCDS.data['dx'] = delta_x
-    residsCDS.data['dy'] = delta_y
-    npoints = len(delta_x)
-    if npoints < 2:
-        return None
-
-    p1 = HAPFigure(title='Residuals for {} [{} sources]: X vs DX'.format(filename, npoints),
-                   x_label="X (pixels)",
-                   y_label='Delta[X] (pixels)',
-                   use_hover_tips=False)
-    p1.build_glyph('circle',
-                   x='x',
-                   y='dx',
-                   sourceCDS=residsCDS)
-
-    p2 = HAPFigure(title='Residuals for {} [{} sources]: X vs DY'.format(filename, npoints),
-                   x_label="X (pixels)",
-                   y_label='Delta[Y] (pixels)',
-                   use_hover_tips=False)
-    p2.build_glyph('circle',
-                   x='x',
-                   y='dy',
-                   sourceCDS=residsCDS)
-
-    row1 = row(p1.fig, p2.fig)
-
-    p3 = HAPFigure(title='Residuals for {} [{} sources]: Y vs DX'.format(filename, npoints),
-                   x_label="Y (pixels)",
-                   y_label='Delta[X] (pixels)',
-                   use_hover_tips=False)
-    p3.build_glyph('circle',
-                   x='y',
-                   y='dx',
-                   sourceCDS=residsCDS)
-
-    p4 = HAPFigure(title='Residuals for {} [{} sources]: Y vs DY'.format(filename, npoints),
-                   x_label="Y (pixels)",
-                   y_label='Delta[Y] (pixels)',
-                   use_hover_tips=False)
-    p4.build_glyph('circle',
-                   x='y',
-                   y='dy',
-                   sourceCDS=residsCDS)
-
-    row2 = row(p3.fig, p4.fig)
-
-    pv = build_vector_plot(residsCDS,
-                           title='Vector plot of Image - Reference for {}'.format(filename),
-                           x_label='X (pixels)',
-                           y_label='Y (pixels)',
-                           color='blue')
-
-    # Display and save
-    if display_plot:
-        show(column(row1, row2, pv.fig))
-    else:
-        print('Saving: {}'.format(output))
-        save(column(row1, row2, pv.fig))
-
-    return output
-
-
-# MDD I only changed this routine so DataFrames are returned from get_pandas_data in this module.
-def build_astrometry_plots(pandas_file,
-                           output_dir=None,
-                           output='cal_qa_results.html',
-                           display_plot=False):
-
-    fit_DF, resids_DF = get_pandas_data(pandas_file)
-    fitCDS = ColumnDataSource(fit_DF)
-    residsCDS = ColumnDataSource(resids_DF)
-
-    if output_dir is not None:
-        summary_filename = os.path.join(output_dir, output)
-    else:
-        summary_filename = output
-
-    # Generate the astrometric plots
-    # astrometry_plot_name = generate_summary_plots(fitCDS, display_plot, output=summary_filename)
-
-    resids_plot_names = []
-
-    for filename, rootname in zip(fitCDS.data[HOVER_COLUMNS[0]], fitCDS.data['index']):
-        i = residsCDS.data['index'].tolist().index(rootname)
-        resids_dict = {'x': residsCDS.data[RESIDS_COLUMNS[0]][i],
-                       'y': residsCDS.data[RESIDS_COLUMNS[1]][i],
-                       'xr': residsCDS.data[RESIDS_COLUMNS[2]][i],
-                       'yr': residsCDS.data[RESIDS_COLUMNS[3]][i]}
-
-        cds = ColumnDataSource(resids_dict)
-
-        resids_plot_name = generate_residual_plots(cds, filename,
-                                                   display_plot,
-                                                   output_dir=output_dir,
-                                                   output=output)
-        resids_plot_names.append(resids_plot_name)
-        break
-
-    resids_plot_names = list(filter(lambda a: a != None, resids_plot_names))
-    return resids_plot_names
